+++ module/config.js
@@ -656,6 +656,7 @@ export function initConfig() {
     ...Object.entries(torgeternity.meleeWeaponTraits).reduce((acc, ent) => { acc[ent[0]] = { label: ent[1], group: 'torgeternity.traitGroup.melee' }; return acc }, {}),
     ...Object.entries(torgeternity.rangedWeaponTraits).reduce((acc, ent) => { acc[ent[0]] = { label: ent[1], group: 'torgeternity.traitGroup.ranged' }; return acc }, {}),
     ...torgeternity.statusEffects.reduce((acc, status) => { acc[status.id] = { label: status.name, group: 'torgeternity.sheetLabels.conditions' }; return acc }, {}),
+    ...Object.entries(torgeternity.skills).reduce((acc, ent) => { acc[ent[0]] = { label: ent[1], group: 'torgeternity.sheetLabels.skills' }; return acc }, {}),
   }