+++ Changelog.md
@@ -61,6 +61,7 @@
 - Allow an Active Effect to be active only when the owning Actor has a particular condition or (equipped) item with a particular trait.
   - For equippable Items (armor, weapons), the item must be equipped for its traits to be considered.
   - e.g. "Cloak of Darkness" effects can have "Active Only with these traits" set to "Dim" (or "Dark" or "Pitch Black") to set the appropriate bonus to toughness.
+- Allow any skill to be specified in an Active Effect's `Apply If Attack/Defend` and `Defend Against` fields.
 
 ### Bug Fixes